Split oversized source files and cut two functions' complexity (SonarCloud)
Clears the SonarCloud new_violations gate (7 findings) on the dev->main
integration PR #545. The gate measures new code against main, so debt that
accumulated on dev across many PRs all reads as new at the integration boundary.

- Split 4 files under the SonarCloud S104 500-line threshold (all were within
  the repo's ADR-015 600-line cap, flagged only by Sonar's stricter default):
  - aces_contracts/participant_behavior.py 599 -> 466 (enums/tables -> _participant_behavior_types.py)
  - aces_sdl/validator/_runtime_platform.py 554 -> 463 (orchestration -> _runtime_orchestration.py)
  - aces_sdl/validator/_relationships.py 525 -> 239 (proxy-upstream -> _relationships_proxy.py)
  - aces_sdl/runtime_datastore_partitions.py 508 -> 354 (node models -> runtime_datastore_nodes.py, helpers -> _runtime_datastore_support.py)
- Reduce ParticipantHistoryViewModel._validate_nested_record_scope cognitive
  complexity 22 -> <=15 (extract the nested walk to module helpers).
- runtime_values.name_indicates_secret: 4 -> 3 returns; drop the S1309-flagged
  #noqa via the file's existing string-split idiom.

No behavior change; public APIs preserved by re-export. Full nox verify passes.
